chore: move getFluxExample to shared folder
